requirements.txt
pyyaml==5.4.1
imageio
opencv-python
moviepy
scikit-image~=0.17.0
gradio~=2.5.0
fastapi
Jinja2
face-alignment